Update Dockerfile
To run docker container, first create an image with: 'docker build -t model .', this looks for the Dockerfile in your current directory and the image takes a while to build the first time. After creating an image you can start a container with 'docker run --entrypoint "/bin/sh" -it model', this opens an interactive shell for you inside the container. From inside the container, you can run 'python model.py'.